Analyzing/src/requirement.txt
paho-mqtt = 1.5.1